[refactor] Hub 포트앤 어댑터 패턴 적용
