+++ hub/src/main/java/sparta/rooibos/hub/application/service/core/HubManagerServiceImpl.java
@@ -7,7 +7,7 @@ import sparta.rooibos.hub.application.dto.hubManager.response.CreateHubManagerRe
 import sparta.rooibos.hub.application.dto.hubManager.response.GetHubManagerHubIdResponse;
 import sparta.rooibos.hub.application.service.exception.BusinessHubManagerException;
 import sparta.rooibos.hub.application.service.exception.custom.HubManagerErrorCode;
-import sparta.rooibos.hub.application.service.port.HubManagerService;
+import sparta.rooibos.hub.application.service.port.in.HubManagerService;
 import sparta.rooibos.hub.domain.model.HubManager;
 import sparta.rooibos.hub.domain.respository.HubManagerRepository;
