+++ hub/src/main/java/sparta/rooibos/hub/application/service/port/in/HubService.java
@@ -1,4 +1,4 @@
-package sparta.rooibos.hub.application.service.port;
+package sparta.rooibos.hub.application.service.port.in;
 
 import sparta.rooibos.hub.application.dto.hub.request.CreateHubRequest;
 import sparta.rooibos.hub.application.dto.hub.request.SearchHubRequest;
